requirements.txt
whitenoise==6.2.0
urllib3==1.26.9
toml==0.10.2
sqlparse==0.4.2
six==1.16.0
s3transfer==0.6.0
pytz==2022.1
python-dateutil==2.8.2
PyJWT==2.4.0
pycodestyle==2.8.0
psycopg2==2.9.3
Pillow==9.2.0
jmespath==1.0.1
gunicorn==20.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-storages==1.13.1
django-cors-headers==3.13.0
Django==4.0
botocore==1.27.46
boto3==1.24.46
autopep8==1.6.0
asgiref==3.4.1